+++ pkg/glib/v2/genviron.go
@@ -27,7 +27,7 @@ func EnvironGetenv(envp []string, variable string) string {
 	var _cret *C.gchar  // in
 
 	{
-		_arg1 = (**C.gchar)(C.malloc(C.ulong(len(envp)+1) * C.ulong(unsafe.Sizeof(uint(0)))))
+		_arg1 = (**C.gchar)(C.malloc(C.size_t(len(envp)+1) * C.size_t(unsafe.Sizeof(uint(0)))))
 		defer C.free(unsafe.Pointer(_arg1))
 		{
 			out := unsafe.Slice(_arg1, len(envp)+1)
@@ -72,7 +72,7 @@ func EnvironSetenv(envp []string, variable, value string, overwrite bool) []stri
 	var _cret **C.gchar  // in
 
 	{
-		_arg1 = (**C.gchar)(C.malloc(C.ulong(len(envp)+1) * C.ulong(unsafe.Sizeof(uint(0)))))
+		_arg1 = (**C.gchar)(C.malloc(C.size_t(len(envp)+1) * C.size_t(unsafe.Sizeof(uint(0)))))
 		{
 			out := unsafe.Slice(_arg1, len(envp)+1)
 			var zero *C.gchar
@@ -132,7 +132,7 @@ func EnvironUnsetenv(envp []string, variable string) []string {
 	var _cret **C.gchar // in
 
 	{
-		_arg1 = (**C.gchar)(C.malloc(C.ulong(len(envp)+1) * C.ulong(unsafe.Sizeof(uint(0)))))
+		_arg1 = (**C.gchar)(C.malloc(C.size_t(len(envp)+1) * C.size_t(unsafe.Sizeof(uint(0)))))
 		{
 			out := unsafe.Slice(_arg1, len(envp)+1)
 			var zero *C.gchar
