+++ pkg/glib/v2/gfileutils.go
@@ -341,7 +341,7 @@ func BuildFilenamev(args []string) string {
 	var _cret *C.gchar  // in
 
 	{
-		_arg1 = (**C.gchar)(C.malloc(C.ulong(len(args)+1) * C.ulong(unsafe.Sizeof(uint(0)))))
+		_arg1 = (**C.gchar)(C.malloc(C.size_t(len(args)+1) * C.size_t(unsafe.Sizeof(uint(0)))))
 		defer C.free(unsafe.Pointer(_arg1))
 		{
 			out := unsafe.Slice(_arg1, len(args)+1)
@@ -382,7 +382,7 @@ func BuildPathv(separator string, args []string) string {
 	_arg1 = (*C.gchar)(unsafe.Pointer(C.CString(separator)))
 	defer C.free(unsafe.Pointer(_arg1))
 	{
-		_arg2 = (**C.gchar)(C.malloc(C.ulong(len(args)+1) * C.ulong(unsafe.Sizeof(uint(0)))))
+		_arg2 = (**C.gchar)(C.malloc(C.size_t(len(args)+1) * C.size_t(unsafe.Sizeof(uint(0)))))
 		defer C.free(unsafe.Pointer(_arg2))
 		{
 			out := unsafe.Slice(_arg2, len(args)+1)
