+++ pkg/graphene/graphene-sphere.go
@@ -249,7 +249,7 @@ func (s *Sphere) InitFromVectors(vectors []Vec3, center *Point3D) *Sphere {
 
 	_arg0 = (*C.graphene_sphere_t)(gextras.StructNative(unsafe.Pointer(s)))
 	_arg1 = (C.uint)(len(vectors))
-	_arg2 = (*C.graphene_vec3_t)(C.malloc(C.ulong(len(vectors)) * C.ulong(C.sizeof_graphene_vec3_t)))
+	_arg2 = (*C.graphene_vec3_t)(C.malloc(C.size_t(len(vectors)) * C.size_t(C.sizeof_graphene_vec3_t)))
 	defer C.free(unsafe.Pointer(_arg2))
 	{
 		out := unsafe.Slice((*C.graphene_vec3_t)(_arg2), len(vectors))
