+++ pkg/gtk/v3/gtkrc.go
@@ -837,7 +837,7 @@ func RCSetDefaultFiles(filenames []string) {
 	var _arg1 **C.gchar // out
 
 	{
-		_arg1 = (**C.gchar)(C.malloc(C.ulong(len(filenames)+1) * C.ulong(unsafe.Sizeof(uint(0)))))
+		_arg1 = (**C.gchar)(C.malloc(C.size_t(len(filenames)+1) * C.size_t(unsafe.Sizeof(uint(0)))))
 		defer C.free(unsafe.Pointer(_arg1))
 		{
 			out := unsafe.Slice(_arg1, len(filenames)+1)
