+++ pkg/gtk/v3/gtktreemodel.go
@@ -1153,7 +1153,7 @@ func (treeModel *TreeModel) RowsReordered(path *TreePath, iter *TreeIter, newOrd
 		_arg2 = (*C.GtkTreeIter)(gextras.StructNative(unsafe.Pointer(iter)))
 	}
 	_arg4 = (C.gint)(len(newOrder))
-	_arg3 = (*C.gint)(C.malloc(C.ulong(len(newOrder)) * C.ulong(C.sizeof_gint)))
+	_arg3 = (*C.gint)(C.malloc(C.size_t(len(newOrder)) * C.size_t(C.sizeof_gint)))
 	defer C.free(unsafe.Pointer(_arg3))
 	{
 		out := unsafe.Slice((*C.gint)(_arg3), len(newOrder))
@@ -1370,7 +1370,7 @@ func NewTreePathFromIndices(indices []int) *TreePath {
 	var _cret *C.GtkTreePath // in
 
 	_arg2 = (C.gsize)(len(indices))
-	_arg1 = (*C.gint)(C.malloc(C.ulong(len(indices)) * C.ulong(C.sizeof_gint)))
+	_arg1 = (*C.gint)(C.malloc(C.size_t(len(indices)) * C.size_t(C.sizeof_gint)))
 	defer C.free(unsafe.Pointer(_arg1))
 	{
 		out := unsafe.Slice((*C.gint)(_arg1), len(indices))
